@@ -28,6 +28,7 @@ namespace icap {
 
 	class Request {
 	public:
+		Request();
 		Request( RequestHeader * req_header );
 		virtual ~Request();
 
@@ -55,14 +56,20 @@ namespace icap {
 		*   Returns the number of preview bytes in the request. If the preview
 		*   header is not present in the request then a minus (-1) value will be
 		*   returned.
-		*   
+		*
 		*   @return preview bytes
 		*/
-	   const int preview_size() throw();	
+		const int preview_size() throw();
+
+		void read( char* buffer, size_t size );
 
 	private:
+		void read_body( char* buf, size_t size );
+
 		RequestHeader * _header;
 		payload_t _payload;
+
+		std::string _data;
 
 	};
 
